+++ src/entities/notification/api/endpoint.ts
@@ -1,0 +1,4 @@
+export const NOTIFICATION_URL = {
+  base: '/notifications',
+  detail: (id: string) => `${NOTIFICATION_URL.base}/${id}`,
+}
