+++ src/entities/notification/api/queryKey.ts
@@ -1,0 +1,4 @@
+export const NOTIFICATION_QUERY_KEY = {
+  all: ['notifications'] as const,
+  read: (id: string) => ['notification', id] as const,
+}
