+++ src/entities/notification/model/type.ts
